+++ src/But/Mpl/ValueWrap.hpp
@@ -1,45 +1,19 @@
 #pragma once
-#include <utility>
-#include <type_traits>
-#include <boost/operators.hpp>
+#include <But/Mpl/detail/ValueWrap.hpp>
 
-namespace But
-{
-namespace Mpl
+namespace But::Mpl
 {
 
-#define BUT_MPL_VALUE_WRAP(name, type) \
-  class name final: public boost::equivalent<name>, \
-                    public boost::totally_ordered<name> \
-  { \
-  public: \
-    template<typename ...Args> \
-    explicit name(Args&& ...args): \
-      /* NOTE: value_(...) not value_{...} - see initializer_list<> overload */ \
-      value_( std::forward<Args>(args)... ) \
-    { } \
-    \
-    template<typename T, \
-             typename = std::enable_if_t< std::is_constructible<type, std::initializer_list<T>>::value > \
-            > \
-    name(std::initializer_list<T> lst): \
-      value_{lst} \
-    { } \
-    \
-    name(name const&) = default; \
-    name& operator=(name const&) = default; \
-    \
-    name(name&&) = default; \
-    name& operator=(name&&) = default; \
-    \
-    auto operator<(name const& rhs) const { return get() < rhs.get(); } \
-    \
-    type const& get() const & { return value_; } \
-    type&& get() && { return std::move(value_); } \
-    \
-  private: \
-    type value_; \
-  }
+/// @brief declares value that's both copyable and movable
+#define BUT_MPL_VALUE_WRAP_CP_MV(name, type) \
+  BUT_MPL_DETAIL_VALUE_WRAP(name, BUT_MPL_DETAIL_VALUE_WRAP_COPYABLE,     BUT_MPL_DETAIL_VALUE_WRAP_MOVABLE,     type)
 
-}
+/// @brief declares value that's copyable but not movable
+#define BUT_MPL_VALUE_WRAP_CP(name, type) \
+  BUT_MPL_DETAIL_VALUE_WRAP(name, BUT_MPL_DETAIL_VALUE_WRAP_COPYABLE,     BUT_MPL_DETAIL_VALUE_WRAP_NON_MOVABLE, type)
+
+/// @brief declares value that's movable but not copyable
+#define BUT_MPL_VALUE_WRAP_MV(name, type) \
+  BUT_MPL_DETAIL_VALUE_WRAP(name, BUT_MPL_DETAIL_VALUE_WRAP_NON_COPYABLE, BUT_MPL_DETAIL_VALUE_WRAP_MOVABLE,     type)
+
 }