+++ src/cmake/flags/clang9.cmake
@@ -1,2 +1,0 @@
-# TODO: temporary, until a nicer mechanism for handling that in BUT_MPL_VALUE_WRAP() macro is found
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-defaulted-function-deleted")